Updated credit to used learning resources
I forgot to include links to the youtube tutorials I followed initially, sorry about that. I found them useful for this fun little project thus far and I think it's important to include them so on the off chance anyone ever sees this text I'm writing, they'll be able to go and do it too.